pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer is a site that analyzes specified pages for SEO suitability, similar to PageSpeed ​​Insights."
authors = ["zhukata <https://github.com/zhukata>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
Flask = "^3.0.3"
flake8 = "^7.1.0"
gunicorn = "^22.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.28.3"
datetime = "^5.5"
requests = "^2.32.3"
bs4 = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
